Death scene
Added a death scene with Replay and Quit options, connected to a death pit and a non-moving enemy. Moved the player into a node2d called PlayerContainer and added a script to that object so that I can apply that script to both the regular player and the mech easily.